refactor(consensus-rewrite-v2): move proof types to lib-proofs::consensus (closes #2329)
Implements CONS-104 per docs/epics/consensus-rewrite-epic.md and AD-003.

Architecture:

- New `lib-proofs/src/consensus/` houses `StakeProof`, `StakeDelegation`,
  `WorkProof`, `ProofOfUsefulWork` plus their self-contained methods
  (`new`, hash-integrity `verify`, `total_work`, `calculate_difficulty`,
  `get_work_score`, and `verify_self_contained`).
- New `lib-consensus/src/proof_verify.rs` keeps the cross-validator chain
  that touches `NetworkState` / `ComputeResult` / lib-storage attestations:
  - free function `verify_pouw_against_network_state(&proof, &network_state)`
    replaces the old `ProofOfUsefulWork::verify(&NetworkState)` method.
  - `impl NetworkState { get_node_routing_work / get_node_storage_proofs /
    get_node_compute_results }` continues to live in lib-consensus
    (orphan rule: `NetworkState` is defined here, simulators must impl here).
- `lib-consensus/src/proofs/` deleted entirely (mod.rs, work_proof.rs).
  `stake_proof.rs` moved via `git mv` so history is preserved.

Why this shape (and not "move everything to lib-proofs"):

- `WorkProof::verify(&NetworkState)` and the network-record simulators
  reference `lib_consensus::types::{NetworkState, ComputeResult}` and
  `lib_storage::proofs::StorageCapacityAttestation`.
- `lib-storage` already depends on `lib-proofs` (for the ZK backend), so
  pulling lib-storage into lib-proofs would create a cycle.
- The Rust orphan rule prevents lib-consensus from adding inherent impls
  on `ProofOfUsefulWork` once the type lives in lib-proofs. Hence the
  free-function rewrite in `lib-consensus/src/proof_verify.rs`.

Dependency direction (CONS-104):

- `lib-proofs::consensus` adds no new external deps. Uses `lib_crypto::Hash`
  directly instead of `lib_identity::IdentityId` (transparent alias) because
  `lib-identity` already depends on `lib-proofs` (would cycle). Defines a
  local `pub type IdentityId = Hash;` for naming compatibility.
- `lib-proofs/Cargo.toml` gains `rand = "0.8"` (used by `WorkProof::new`
  for nonce generation, previously inherited via lib-consensus's deps).
- `lib-consensus/src/lib.rs` re-exports the moved types via
  `pub use lib_proofs::consensus::{...};` for callers that still use
  `lib_consensus::{StakeProof, WorkProof, ...}` paths during the rewrite.

Engine wiring:

- `lib-consensus/src/engines/consensus_engine/proofs.rs`:
  `use crate::proofs::StorageCapacityAttestation;` →
  `use lib_storage::proofs::StorageCapacityAttestation;` (the type was
  always a lib-storage type; the lib-consensus path was just a re-export).
- `lib-consensus/src/engines/consensus_engine/mod.rs`:
  `crate::proofs::StorageProofProvider` → `lib_storage::proofs::StorageProofProvider`
  (same: lib-storage type, was re-exported).
- `lib-consensus/src/types/mod.rs` re-exports updated to source from
  `lib_proofs::consensus` and `lib_storage::proofs` directly.

Test coverage:

- `lib-consensus/tests/proof_tests.rs:271`: replaces
  `proof.verify(&network_state)` with
  `lib_consensus::proof_verify::verify_pouw_against_network_state(&proof, &network_state)`.
- All 16 proof_tests still pass.
- All other lib-consensus tests preserve the KI-01..KI-07 baseline exactly.
- `cargo build --workspace` clean. `cargo test -p lib-proofs` passes.

Closes #2329
Refs #2365
