+++ lib-consensus/tests/proof_tests.rs
@@ -268,7 +268,13 @@ fn test_proof_of_useful_work_verification() -> Result<()> {
         node_id,
     )?;
 
-    let is_valid = proof.verify(&network_state)?;
+    // CONS-104: ProofOfUsefulWork::verify(&NetworkState) was extracted to a free
+    // function in lib-consensus because the orphan rule blocks adding inherent
+    // impls on the type (now in lib-proofs) from this crate.
+    let is_valid = lib_consensus::proof_verify::verify_pouw_against_network_state(
+        &proof,
+        &network_state,
+    )?;
 
     // Should be valid for test network state
     assert!(is_valid);
